Allow setting of VM hardware version
This change adds a new model level flag, that allows operators to
set a newer compatibility version for the instances that get spawned
by juju.

The version set via model config must be of equal or higher version
than that of the source template, and it must be supported by the
remote compute resource it will be spawned on. Attempting to set a
version that does not conform to the mentioned rules, will be ignored,
and the VM will use the same HW compatibility version as the source
template.